% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriROC.R
\name{PriROC}
\alias{PriROC}
\title{Compute differentially private ROC curve}
\usage{
PriROC(X, epsilon, ratio = 0.2)
}
\arguments{
\item{X}{n by 2 matrix, with row_i = (L_i,P_i) for the label and prediction of record_i}

\item{epsilon}{privacy budget}

\item{ratio}{budget epsilon*ratio spent on selecting thresholds, default is 0.2}
}
\value{
A list containing nAUC for noisy AUC and nROC for noisy ROC curve of (TPR,FPR) points
}
\description{
Compute differentially private ROC curve
}
\examples{
library(VerificationMeasures)
library(Matrix)
library(smoothmest)
options(expressions = 10000)
example.file <- system.file("extdata", "ratios_twitter.csv", package = "VerificationMeasures")
data <- read.csv(example.file,header = FALSE)
X <- as.matrix(data[,c(1,2)]) #only use the first two columns?
epsilon <- 1
result <- PriROC(X,epsilon)
FPR <- result$nROC[,2]
TPR <- result$nROC[,1]
#plot(FPR, TPR, type = 'l', xlab = "False Positive Rate", ylab = "True Positive Rate")
}

